Update and rename test_numeric.py to test_math.py
